+++ memoria/src/memoria_de_usuario.h
@@ -17,8 +17,8 @@
 #include <memoria/memoria.h>
 #include <stdint.h>
 
-uint8_t leer_de_memoria(t_direccion_fisica* direccion);
-void escribir_en_memoria(t_direccion_fisica* direccion, uint8_t a_escribir);
+uint32_t leer_de_memoria(t_direccion_fisica* direccion);
+void escribir_en_memoria(t_direccion_fisica* direccion, uint32_t a_escribir);
 int32_t obtener_numero_de_marco(uint32_t pid, uint32_t pagina_buscada);
 void asignar_memoria(uint32_t pid, uint32_t size, uint32_t (*algoritmo)(void), t_list* bloques_swap);
 uint32_t buscar_victima_fifo(void);
